+++ doc/changelog.rst
@@ -4,6 +4,7 @@ Release Notes
 =============
 
 - :release:`unreleased <????-??-??>`
+- :feature:`1254` Organisers can now also send emails to reviewers or other team members.
 - :bug:`-` An organiser-specific help text on room availability forms would show up for speaker availabilities, too.
 - :bug:`-` Review scores and phases were not copied correctly when copying event settings.
 - :bug:`-` Sessions could not be scheduled for midnight via drag'n'drop (though scheduling via the session form still worked as a workaround).